Refactoring

+++ database/code/src/ClientManager.cpp
@@ -9,14 +9,8 @@ ClientManager::ClientManager(client _my_client): table_name("clients"),
                 my_client(std::move(_my_client)) {
 }
 
-bool ClientManager::add(std::string data) {
-    if (!dbManager.is_db_exists()) {
-        dbManager.create_db();
-    }
-
-    if (!dbManager.is_connected_to_db()) {
-        dbManager.connect_to_db();
-    }
+bool ClientManager::add() {
+    check_access();
 
     if (!dbManager.is_table_exists(table_name)) {
         std::cout << "TABLE NOT EXIST\n";
@@ -31,6 +25,7 @@ bool ClientManager::add(std::string data) {
 }
 
 std::vector<std::vector<std::string>> ClientManager::get(int id) {
+    check_access();
     std::vector<std::vector<std::string>> to_return = dbManager.get_data(
          table_name, my_client.get_cols()
     );
@@ -39,6 +34,7 @@ std::vector<std::vector<std::string>> ClientManager::get(int id) {
 }
 
 bool ClientManager::erase(int id) {
+    check_access();
     return dbManager.delete_data(table_name, id);
 }
 
@@ -55,6 +51,11 @@ std::vector<std::string> ClientManager::to_vector() {
 }
 
 bool ClientManager::update(int id, std::vector<std::pair<std::string, std::string>> &val) {
+    check_access();
+    return dbManager.update_data(table_name, val, id);
+}
+
+void ClientManager::check_access() {
     if (!dbManager.is_db_exists()) {
         dbManager.create_db();
     }
@@ -62,6 +63,4 @@ bool ClientManager::update(int id, std::vector<std::pair<std::string, std::strin
     if (!dbManager.is_connected_to_db()) {
         dbManager.connect_to_db();
     }
-
-    return dbManager.update_data(table_name, val, id);
 }
